A file name must no longer be specified
The program now utilizes the datetime library's function  datetime.now() function to generate a new file name. This is useful especially when accessing a current bullet journal file because if that day is the same day as the file name it does not need to be a command line argument.